pyproject.toml
[tool.poetry]
name = "films"
version = "0.1.0"
description = ""
authors = ["FrozZzy1 <frozzzy71@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.2"
databases = "^0.8.0"
uvicorn = "^0.23.2"
asyncpg = "^0.28.0"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
pydantic = {extras = ["email"], version = "^2.4.2"}
sqlalchemy-utils = "^0.41.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
